Support other networks than Mainnet for Sushiswap
- Organize Sushiswap and Kyber oracles files
- Related to #343
