+++ gnosis/eth/oracles/oracles.py
@@ -18,7 +18,6 @@ from .. import EthereumClient, EthereumNetwork
 from ..constants import NULL_ADDRESS
 from ..contracts import (
     get_erc20_contract,
-    get_kyber_network_proxy_contract,
     get_uniswap_factory_contract,
     get_uniswap_v2_factory_contract,
     get_uniswap_v2_pair_contract,
@@ -100,89 +99,6 @@ def get_decimals(
         error_message = f"Cannot get decimals for token={token_address}"
         logger.warning(error_message)
         raise CannotGetPriceFromOracle(error_message) from e
-
-
-class KyberOracle(PriceOracle):
-    # This is the `tokenAddress` they use for ETH ¯\_(ツ)_/¯
-    ETH_TOKEN_ADDRESS = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"
-    ADDRESSES = {
-        EthereumNetwork.MAINNET: "0x9AAb3f75489902f3a48495025729a0AF77d4b11e",
-        EthereumNetwork.RINKEBY: "0x0d5371e5EE23dec7DF251A8957279629aa79E9C5",
-        EthereumNetwork.ROPSTEN: "0xd719c34261e099Fdb33030ac8909d5788D3039C4",
-        EthereumNetwork.KOVAN: "0xc153eeAD19e0DBbDb3462Dcc2B703cC6D738A37c",
-    }
-
-    def __init__(
-        self,
-        ethereum_client: EthereumClient,
-        kyber_network_proxy_address: Optional[str] = None,
-    ):
-        """
-        :param ethereum_client:
-        :param kyber_network_proxy_address: https://developer.kyber.network/docs/MainnetEnvGuide/#contract-addresses
-        """
-        self.ethereum_client = ethereum_client
-        self.w3 = ethereum_client.w3
-        self._kyber_network_proxy_address = kyber_network_proxy_address
-
-    @cached_property
-    def kyber_network_proxy_address(self):
-        if self._kyber_network_proxy_address:
-            return self._kyber_network_proxy_address
-        return self.ADDRESSES.get(
-            self.ethereum_client.get_network(),
-            self.ADDRESSES.get(EthereumNetwork.MAINNET),
-        )  # By default return Mainnet address
-
-    @cached_property
-    def kyber_network_proxy_contract(self):
-        return get_kyber_network_proxy_contract(
-            self.w3, self.kyber_network_proxy_address
-        )
-
-    def get_price(
-        self, token_address_1: str, token_address_2: str = ETH_TOKEN_ADDRESS
-    ) -> float:
-        if token_address_1 == token_address_2:
-            return 1.0
-        try:
-            # Get decimals for token, estimation will be more accurate
-            decimals = self.ethereum_client.erc20.get_decimals(token_address_1)
-            token_unit = int(10**decimals)
-            (
-                expected_rate,
-                _,
-            ) = self.kyber_network_proxy_contract.functions.getExpectedRate(
-                token_address_1, token_address_2, int(token_unit)
-            ).call()
-
-            price = expected_rate / 1e18
-
-            if price <= 0.0:
-                # Try again the opposite
-                (
-                    expected_rate,
-                    _,
-                ) = self.kyber_network_proxy_contract.functions.getExpectedRate(
-                    token_address_2, token_address_1, int(token_unit)
-                ).call()
-                price = (token_unit / expected_rate) if expected_rate else 0
-
-            if price <= 0.0:
-                error_message = (
-                    f"price={price} <= 0 from kyber-network-proxy={self.kyber_network_proxy_address} "
-                    f"for token-1={token_address_1} to token-2={token_address_2}"
-                )
-                logger.warning(error_message)
-                raise InvalidPriceFromOracle(error_message)
-            return price
-        except (ValueError, BadFunctionCallOutput, DecodingError) as e:
-            error_message = (
-                f"Cannot get price from kyber-network-proxy={self.kyber_network_proxy_address} "
-                f"for token-1={token_address_1} to token-2={token_address_2}"
-            )
-            logger.warning(error_message)
-            raise CannotGetPriceFromOracle(error_message) from e
 
 
 class UniswapOracle(PriceOracle):
@@ -324,12 +240,13 @@ class UniswapOracle(PriceOracle):
 
 
 class UniswapV2Oracle(PricePoolOracle, PriceOracle):
+    ROUTER_ADDRESSES = {
+        EthereumNetwork.MAINNET: "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
+    }
+
     # Pair init code is keccak(getCode(UniswapV2Pair))
-    pair_init_code = HexBytes(
+    PAIR_INIT_CODE = HexBytes(
         "0x96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f"
-    )
-    router_address = (
-        "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D"  # Same address on every network
     )
 
     def __init__(
@@ -341,7 +258,10 @@ class UniswapV2Oracle(PricePoolOracle, PriceOracle):
         """
         self.ethereum_client = ethereum_client
         self.w3 = ethereum_client.w3
-        self.router_address: str = router_address or self.router_address
+        self.router_address: str = router_address or self.ROUTER_ADDRESSES.get(
+            ethereum_client.get_network(),
+            self.ROUTER_ADDRESSES[EthereumNetwork.MAINNET],
+        )
         self.router = get_uniswap_v2_router_contract(
             ethereum_client.w3, self.router_address
         )
@@ -407,7 +327,7 @@ class UniswapV2Oracle(PricePoolOracle, PriceOracle):
         address = fast_keccak(
             encode_abi_packed(
                 ["bytes", "address", "bytes", "bytes"],
-                [HexBytes("ff"), self.factory_address, salt, self.pair_init_code],
+                [HexBytes("ff"), self.factory_address, salt, self.PAIR_INIT_CODE],
             )
         )[-20:]
         return fast_bytes_to_checksum_address(address)
@@ -538,13 +458,6 @@ class UniswapV2Oracle(PricePoolOracle, PriceOracle):
             )
             logger.warning(error_message)
             raise CannotGetPriceFromOracle(error_message) from e
-
-
-class SushiswapOracle(UniswapV2Oracle):
-    pair_init_code = HexBytes(
-        "0xe18a34eb0e04b04f7a0ac29a6e80748dca96319b42c54d679cb821dca90c6303"
-    )
-    router_address = "0xd9e1cE17f2641f24aE83637ab66a2cca9C378B9F"
 
 
 class AaveOracle(PriceOracle):
